+++ src/main/java/com/github/kd_gaming1/skyblockenhancements/repo/neu/NeuItem.java
@@ -192,5 +192,6 @@ public class NeuItem {
             this.recipesTrimmed = true;
         }
         this.recipes = null;
+        this.recipe = null;
     }
 }